project_asistente_de_formulario_de_ventas_anterior/requirements.txt
google-adk
pygsheets
pandas
python-dotenv